requirements.txt
# LeetCode Discord Bot - Python Dependencies

# Discord.py library for bot functionality
discord.py>=2.3.0

# Async HTTP client for API requests
aiohttp>=3.8.0

# Async SQLite support
aiosqlite>=0.19.0

# Environment variable management
python-dotenv>=1.0.0
